Renomeação da biblioteca
